+++ src/main/kotlin/kz/qpexpress/qpexpress/model/Payment.kt
@@ -6,14 +6,14 @@ import java.time.LocalDateTime
 
 @Entity
 class KaspiPayment(): AbstractJpaPersistable() {
-    var paymentId: Int? = null
+    var paymentId: Long? = null
     var totalAmount: Double? = null
     var availableReturnAmount: Double? = null
     var date: LocalDateTime? = null
     @Enumerated(EnumType.STRING)
     var status: PaymentStatus = PaymentStatus.CREATING
 
-    constructor(paymentId: Int, totalAmount: Double, availableReturnAmount: Double, date: LocalDateTime, status: PaymentStatus): this() {
+    constructor(paymentId: Long, totalAmount: Double, availableReturnAmount: Double, date: LocalDateTime, status: PaymentStatus): this() {
         this.paymentId = paymentId
         this.totalAmount = totalAmount
         this.availableReturnAmount = availableReturnAmount
