+++ src/main/kotlin/kz/qpexpress/qpexpress/repository/KaspiPaymentRepository.kt
@@ -4,5 +4,5 @@ import kz.qpexpress.qpexpress.model.KaspiPayment
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface KaspiPaymentRepository: JpaRepository<KaspiPayment, Int>{
-    fun findByPaymentId(paymentId: Int): KaspiPayment?
+    fun findByPaymentId(paymentId: Long): KaspiPayment?
 }
